+++ ui/helpers/higher-order-components/authenticated/authenticated.component.js
@@ -10,12 +10,23 @@ import {
 export default function Authenticated(props) {
   const { isUnlocked, completedOnboarding } = props;
   switch (true) {
+    // For ONBOARDING_V2 dev purposes,
+    // Remove when ONBOARDING_V2 dev complete
     case process.env.ONBOARDING_V2 === true:
       return <Redirect to={{ pathname: ONBOARDING_ROUTE }} />;
+
     case isUnlocked && completedOnboarding:
       return <Route {...props} />;
     case !completedOnboarding:
-      return <Redirect to={{ pathname: INITIALIZE_ROUTE }} />;
+      return (
+        <Redirect
+          to={{
+            pathname: process.env.ONBOARDING_V2
+              ? ONBOARDING_ROUTE
+              : INITIALIZE_ROUTE,
+          }}
+        />
+      );
     default:
       return <Redirect to={{ pathname: UNLOCK_ROUTE }} />;
   }
