+++ ui/helpers/higher-order-components/initialized/initialized.component.js
@@ -1,13 +1,19 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Redirect, Route } from 'react-router-dom';
-import { INITIALIZE_ROUTE } from '../../constants/routes';
+import { INITIALIZE_ROUTE, ONBOARDING_ROUTE } from '../../constants/routes';
 
 export default function Initialized(props) {
   return props.completedOnboarding ? (
     <Route {...props} />
   ) : (
-    <Redirect to={{ pathname: INITIALIZE_ROUTE }} />
+    <Redirect
+      to={{
+        pathname: process.env.ONBOARDING_V2
+          ? ONBOARDING_ROUTE
+          : INITIALIZE_ROUTE,
+      }}
+    />
   );
 }
