+++ ui/selectors/selectors.js
@@ -171,21 +171,6 @@ export const getMetaMaskAccounts = createSelector(
     ),
 );
 
-export function getCompletedOnboarding(state) {
-  return state.metamask.completedOnboarding;
-}
-export function getIsInitialized(state) {
-  return state.metamask.isInitialized;
-}
-
-export function getIsUnlocked(state) {
-  return state.metamask.isUnlocked;
-}
-
-export function getSeedPhraseBackedUp(state) {
-  return state.metamask.seedPhraseBackedUp;
-}
-
 export function getSelectedAddress(state) {
   return state.metamask.selectedAddress;
 }
